% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_frequency_minute.R
\name{stat_frequency_minute}
\alias{stat_frequency_minute}
\title{Get the physical flow from the statnett api.}
\usage{
stat_frequency_minute(from_date, date_as_cet = FALSE)
}
\arguments{
\item{from_date}{Class Date. From date.}

\item{date_as_cet}{Boolean. Default false. Return the date as CET or not.}
}
\description{
The function returns the datetime as UTC as default.
}
