% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_physical_flow_map.R
\name{stat_physical_flow_map}
\alias{stat_physical_flow_map}
\title{Get the physical flow map data from the statnett api.}
\usage{
stat_physical_flow_map(date_as_cet = FALSE)
}
\arguments{
\item{date_as_cet}{Boolean. Default false. Return the date as CET or not.}
}
\description{
The function returns the datetime as UTC as default.
}
