% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_primary_reserves_day.R
\name{stat_primary_reserves_day}
\alias{stat_primary_reserves_day}
\title{Get the physical flow from the statnett api.}
\usage{
stat_primary_reserves_day(local_date)
}
\arguments{
\item{from_date}{Class Date. From date.}

\item{date_as_cet}{Boolean. Default false. Return the date as CET or not.}
}
\description{
The function returns the datetime as UTC as default.
}
