% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_production_consumption.R
\name{stat_production_consumption}
\alias{stat_production_consumption}
\title{Get the production and consumption data from the statnett api.}
\usage{
stat_production_consumption(from_date, date_as_cet = FALSE)
}
\arguments{
\item{from_date}{Class Date. From date.}

\item{date_as_cet}{Boolean. Default false. Return the date as CET or not.}
}
\description{
The function returns the datetime as UTC as default.
}
