% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_production_consumption_latests_overview.R
\name{stat_production_consumption_latest_overview}
\alias{stat_production_consumption_latest_overview}
\title{Get the latest detailed overview statnett api.}
\usage{
stat_production_consumption_latest_overview(date_as_cet = FALSE,
  data_wide = FALSE)
}
\arguments{
\item{date_as_cet}{Boolean. Default false. Return the date as CET or not.}
}
\description{
The function returns the datetime as UTC as default.
}
